Removed redudancies - qualifiers - arguments - initializers - empty declarations and statements - elaborated type specifiers - conditional expressions - access specifiers - casts

+++ src/common/crypto/CHuffman.cpp
@@ -88,7 +88,7 @@ uint CHuffman::Compress( byte * pOutput, const byte * pInput, uint outLen, uint 
             return 0;
 
         // pOutput[iLen++] = xOutVal << (8u - bitidx);
-		pOutput[iLen++] = (byte)((uint)xOutVal << (uint)(8u - bitidx));
+		pOutput[iLen++] = (byte)((uint)xOutVal << (8u - bitidx));
 	}
 
 	return iLen;
